refactor(desktop): store types extraction, initial state dedup
