+++ apps/desktop/src/store/slices/session-view/createInitialSessionViewState.ts
@@ -1,0 +1,11 @@
+type Params = Record<string, never>;
+
+export const createInitialSessionViewState = ({}: Params) => ({
+  sessionViewPrefs: {},
+  activeLens: {},
+  lensHistory: {},
+  focusedPlanId: {},
+  sessionStudio: {},
+  workflowExpand: {},
+  focusedWorkflowRunId: {},
+});
