+++ apps/desktop/src/store/slices/sessions/setCurrentSession.ts
@@ -20,7 +20,7 @@ import { listPlansForSession as invokeListPlansForSession } from '../../../featu
 import type { AgentKind } from '../../../features/session/agent-kind';
 import { SETTING_LAST_SESSION_ID } from '../../../features/settings/settings';
 import { EMPTY_LOADING } from '../../session-mutators';
-import type { SessionLoadingFlags } from '../../store';
+import type { SessionLoadingFlags } from '../../types';
 import type { GetFn, SetFn } from './types';
 
 export const setCurrentSession = (set: SetFn, get: GetFn) => {
