+++ apps/desktop/src/store/transcript.ts
@@ -1,6 +1,7 @@
 import { useRef } from 'react';
 import type { AgentId, TurnEvent } from '@goodboy/types';
-import { useAppStore, type AppState } from './store';
+import { useAppStore } from './store';
+import type { AppState } from './types';
 
 const EMPTY: ReadonlyArray<TurnEvent> = [];
